requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.23.0
transformers>=4.30.0
accelerate>=0.20.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
huggingface-hub>=0.16.0
xformers>=0.0.20 